Updating to UFR format
Added build script and preprocessing script to adjust the UFOs to be GF compliant versions (especially since the italic position is halfway on the axis that the source files contain).

Tried to maintain sources as-is, but had to change the version string values to match what is in the name table to keep alignment with how fontmake builds the font.